this script also uses the same bus that was created in the i2c_bus.py file.
the function detectarCruce is not needed.

big changes to the stateManager function.
it does not convert angles to its x,y components anymore.
for that reason, there is no ListaTriggersxy.
there is no "hysteresys".
triggers are now always fixed at the same angle.

some variables are now persistent and are included as parameters of the function stateManager():
pendientePos
zona_pasada

No need for the function detectarCruce
rather than detecting crossing over a trigger using x,y: we now check for zones.
pendientePos now depends on the change in zones...

def leer_brillo() and guardar_brillo() are used to work with the new config.json file.

send_command() is used instead of init_display() when setting the display color.

brightness is stored in a file. and is set by adjusting the value and waiting for 5 seconds.